Fix nested escaping of `\` escape symbol (#66)
Essentially it is an additional set of fixes, for the previous PR https://github.com/TykTechnologies/graphql-go-tools/pull/61
Quoting and Unquoting is not that straightforward, take a loot at https://golang.org/src/encoding/json/decode.go `unquoteBytes` function. 

We continue hitting bugs related to escaped symbols, because built-in "escape" function does not handle some cases. And based on the implementation I see in Golang there are way more. 

This PR makes it use standard `strconv.Quote` function, which handles all these scenarios. 

Additionally, I was not able to figure out what this lines do (and frankly the logic itself is a bit off), so I just removed them:
https://github.com/TykTechnologies/graphql-go-tools/pull/66/files#diff-b07a4e25e3d7f6482ed988bdf67cbfe8L280-L281

Plus in some test we dependent on the byte diff viewer which requires goland to be installed, I made it work with `assert` package, and it actually prints a really nice bytes diff. 

Fix https://github.com/TykTechnologies/graphql-go-tools/issues/65